// Practica6/frontend/yo-usac-frontend/src/types/grabaciones.types.ts
export interface Capitulo {
  id_capitulo: number
  id_clase: number
  titulo: string
  tiempo_inicio: number
}